void_installer: config file loading, basic installer structure
each module should be a python file that contains a function `setup()`
and a subclass of `Module`. This will allow them to each define hooks
that the installer will run in each phase of the installation.
Configuration is also handled by the class-within-a-class using pydantic
to convert a dict (from yaml) into a useful object.

this commit adds dependencies on pyyaml and pydantic. although
dependencies were hoped to be avoided in this project, these two are too
useful to live without.